+++ internal/scannergen/backend/golangtable/doc.go
@@ -1,0 +1,3 @@
+// Package golangtable contains the backend which writes the scanner as Go source code holding the automaton in lookup
+// tables.
+package golangtable