+++ pkg/scannergen/backend/golangtable/doc.go
@@ -1,0 +1,3 @@
+// Package golangtable provides a scanner generator backend which takes a DFA as input and outputs Go source code
+// holding the automaton in lookup tables.
+package golangtable